### 2020-03-04 #2139 @whedon
[PRE REVIEW]: Surprise: A Python library for recommender systems
**Submitting author:** @NicolasHug (<a href="http://orcid.org/0000-0003-1360-704X">Nicolas Hug</a>)
**Repository:** <a href="https://github.com/NicolasHug/Surprise" target ="_blank">https://github.com/NicolasHug/Surprise</a>
**Version:** v1.1.0
