### 2020-03-04 #2139 @NicolasHug
Hi @terrytangyuan ,

From the list, sara-02, mhetrerajat, and mamrehn seem to have experience with recommender systems and Python.
