### 2020-03-04 #2139 @arfon
The issue here is that this is not a URL that Git can clone: https://github.com/NicolasHug/Surprise/tree/joss_paper . I've updated the URL here and in the long run we will have Whedon check for this but for now you need to do this 👇
