### 2020-03-19 #2139 @arfon
> I just saw the notification. Is this still open for review?

@sara-02 - yes, we're still processing submissions, albeit at a slower pace. I'll defer to the handling editor (@terrytangyuan) here but I think we'd likely still want you to review and mid-April sounds pretty reasonable to me :-)
