### 2020-03-19 #2139 @arfon
> @arfon Do we still need at least two reviewers at this time?

I would prefer so yes but this might be worth raising with the other JOSS editors on the mailing list?
