### 2020-04-13 #2174 @ejhigson
Hi @NicolasHug - congratulations on making a really awesome and very popular software package!

The install and examples all worked perfectly for me. The API documentation is very thorough too. I have couple of questions:

1) are there any similar other existing software packages for recommender systems? One of the review criteria is "Do the authors describe how this software compares to other commonly-used packages?" and I don't see any such comparison in your paper, but if no such packages exist then this is fine.

2) are there any instructions on how to run the unit tests? I can't see any such instructions in the documentation (sorry if I have missed them). When I try running the tests with ``nosetests`` I get an import error for some module called 'similarities' on every test - I include an example below. Do you know what I need to do to fix this?
```
$ nosetests
EEEEEEEEEEEEEEEEEE
ERROR: Failure: ImportError (cannot import name 'similarities')
Traceback (most recent call last):
  File "/home/ed/.local/lib/python3.6/site-packages/nose/failure.py", line 39, in runTest
    raise self.exc_val.with_traceback(self.tb)
  File "/home/ed/.local/lib/python3.6/site-packages/nose/loader.py", line 418, in loadTestsFromName
    addr.filename, addr.module)
  File "/home/ed/.local/lib/python3.6/site-packages/nose/importer.py", line 47, in importFromPath
    return self.importFromDir(dir_path, fqname)
  File "/home/ed/.local/lib/python3.6/site-packages/nose/importer.py", line 94, in importFromDir
    mod = load_module(part_fqname, fh, filename, desc)
  File "/usr/lib/python3.6/imp.py", line 245, in load_module
    return load_package(name, filename)
  File "/usr/lib/python3.6/imp.py", line 217, in load_package
    return _load(spec)
  File "<frozen importlib._bootstrap>", line 684, in _load
  File "<frozen importlib._bootstrap>", line 665, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 678, in exec_module
  File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
  File "/home/ed/code/surprise/surprise/__init__.py", line 3, in <module>
    from .prediction_algorithms import AlgoBase
  File "/home/ed/code/surprise/surprise/prediction_algorithms/__init__.py", line 23, in <module>
    from .algo_base import AlgoBase
  File "/home/ed/code/surprise/surprise/prediction_algorithms/algo_base.py", line 10, in <module>
    from .. import similarities as sims
ImportError: cannot import name 'similarities'
```
