### 2020-04-13 e054469
Close #328, set precision and recall edge cases to 0
