### 2020-04-13 2ed7709
Add a note on precision_recall_at_k edge cases
