### 2020-04-14 #2174 @sara-02
@NicolasHug If not an exhaustive comparison then even an overview would do. I assume most projects on Recommendation Systems have been written as part of research projects and not maintained after that.
I recently found this markdown that is trying to curate a list of RecSys available for use, maybe you this to curate your related work section of the paper.
https://github.com/grahamjenson/list_of_recommender_systems
