### 2020-04-14 3eded91
Apply suggestions from code review
