+++ examples/precision_recall_at_k.py
@@ -37,11 +37,13 @@ def precision_recall_at_k(predictions, k=10, threshold=3.5):
                               for (est, true_r) in user_ratings[:k])
 
         # Precision@K: Proportion of recommended items that are relevant
-        # When n_rec_k is 0, Precision is undefined. As an alternative, it can be set to 0
+        # When n_rec_k is 0, Precision is undefined. We here set it to 0.
+
         precisions[uid] = n_rel_and_rec_k / n_rec_k if n_rec_k != 0 else 0
 
         # Recall@K: Proportion of relevant items that are recommended
-        # When n_rel is 0, Recall is undefined. As an alternative, it can be set to 0
+        # When n_rel is 0, Recall is undefined. We here set it to 0.
+
         recalls[uid] = n_rel_and_rec_k / n_rel if n_rel != 0 else 0
 
     return precisions, recalls
