### 2020-04-14 3cdf617
Set precision and recall to 0 when undefined (#349)
