### 2020-04-14 #2174 @ejhigson
@NicolasHug :
> I didn't include a comparison with other similar packages. Should I?

I agree with @sara-02 that an overview would be useful.

> recommendation has a very wide range of applications and usually, each library targets a specific one. That I know of, surprise is the only library that focuses on rating prediction and that is still maintained, in Python. There are also openrec, Spotlight and LightFM but they target different applications.

I think this is useful information to include in the paper.

> There are instructions to run the tests in the contributing guidelines: https://github.com/NicolasHug/Surprise/blob/master/CONTRIBUTING.md . Also note that a CI instance is ran at every commit, there's a "build passing badge in the [readme](https://github.com/NicolasHug/Surprise).

Aha sorry I missed this! With these instructions all the tests work perfectly for me.
