### 2020-05-20 #2174 @arfon
:wave: @NicolasHug - today we [reopened JOSS](https://blog.joss.theoj.org/2020/05/reopening-joss) for new submissions and are checking in on our existing reviews. How are you getting on making these updates to your paper?
