### 2020-06-04 2d91d98
Fix FilexExistsError due to race condition (#347)
