### 2020-07-14 1c4d4b9
Fix FilexExistsError due to race condition (#359)
