+++ paper.bib
@@ -91,3 +91,57 @@
   year={2016},
   publisher={Springer}
 }
+
+@inproceedings{librec,
+  title={LibRec: A Java Library for Recommender Systems.},
+  author={Guo, Guibing and Zhang, Jie and Sun, Zhu and Yorke-Smith, Neil}
+}
+
+@inproceedings{mymedialite,
+    author    = { Zeno Gantner and Steffen Rendle and Christoph Freudenthaler and Lars Schmidt-Thieme },
+    title     = { {MyMediaLite}: A Free Recommender System Library },
+    booktitle = { Proceedings of the 5th ACM Conference
+                  on Recommender Systems (RecSys 2011) },
+    year      = 2011
+}
+
+@misc{spotlight,
+  title={Spotlight},
+  author={Kula, Maciej},
+  year={2017},
+  publisher={GitHub},
+  howpublished={\url{https://github.com/maciejkula/spotlight}},
+}
+
+@inproceedings{openrec,
+author = {Yang, Longqi and Bagdasaryan, Eugene and Gruenstein, Joshua and Hsieh, Cheng-Kang and Estrin, Deborah},
+title = {OpenRec: A Modular Framework for Extensible and Adaptable Recommendation Algorithms},
+year = {2018},
+isbn = {9781450355810},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3159652.3159681},
+doi = {10.1145/3159652.3159681},
+booktitle = {Proceedings of the Eleventh ACM International Conference on Web Search and Data Mining},
+pages = {664–672},
+numpages = {9},
+keywords = {recommendation, adaptable, modular, extensible, framework},
+location = {Marina Del Rey, CA, USA},
+series = {WSDM ’18}
+}
+
+@inproceedings{lightfm,
+  author    = {Maciej Kula},
+  editor    = {Toine Bogers and
+               Marijn Koolen},
+  title     = {Metadata Embeddings for User and Item Cold-start Recommendations},
+  booktitle = {Proceedings of the 2nd Workshop on New Trends on Content-Based Recommender
+               Systems co-located with 9th {ACM} Conference on Recommender Systems
+               (RecSys 2015), Vienna, Austria, September 16-20, 2015.},
+  series    = {{CEUR} Workshop Proceedings},
+  volume    = {1448},
+  pages     = {14--21},
+  publisher = {CEUR-WS.org},
+  year      = {2015},
+  url       = {http://ceur-ws.org/Vol-1448/paper4.pdf},
+}
