### 2020-07-14 #2174 @NicolasHug
Thanks again everyone! I have updated the paper with a section on related libraries, as suggested by @sara-02 .
