### 2020-07-17 #2174 @sara-02
@NicolasHug Minor suggestion for the first line for the first line of the paper can we replace the phrase `items that the system has to offer` to `items that the platform/service has to offer` so as not to confuse the word `system` from the RecSys we refer to in the beginning of the sentence.
