### 2020-07-17 93f3f8e
minor adjustments to paper
