### 2020-07-19 beb47bd
minor readme typo edits (#355)
