+++ examples/precision_recall_at_k.py
@@ -12,7 +12,7 @@ from surprise.model_selection import KFold
 
 
 def precision_recall_at_k(predictions, k=10, threshold=3.5):
-    '''Return precision and recall at k metrics for each user.'''
+    """Return precision and recall at k metrics for each user"""
 
     # First map the predictions to each user.
     user_est_true = defaultdict(list)
