### 2020-07-19 1aef5d9
Unify docstrings: single quotes to double quotes (#304)
Co-authored-by: feger <marc.feger@hhu.de>
