+++ surprise/utils.py
@@ -1,4 +1,4 @@
-'''The utils module contains the get_rng function.'''
+"""The utils module contains the get_rng function."""
 
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
@@ -8,12 +8,12 @@ import numpy as np
 
 
 def get_rng(random_state):
-    '''Return a 'validated' RNG.
+    """Return a 'validated' RNG.
 
     If random_state is None, use RandomState singleton from numpy.  Else if
     it's an integer, consider it's a seed and initialized an rng with that
     seed. If it's already an rng, return it.
-    '''
+    """
     if random_state is None:
         return np.random.mtrand._rand
     elif isinstance(random_state, (numbers.Integral, np.integer)):
