+++ paper.bib
@@ -58,6 +58,7 @@
   title     = { Data Structures for Statistical Computing in Python },
   booktitle = { Proceedings of the 9th Python in Science Conference },
   pages     = { 51 - 56 },
+  doi       = {10.25080/Majora-92bf1922-00a},
   year      = { 2010 },
   editor    = { St\'efan van der Walt and Jarrod Millman }
 }
@@ -69,6 +70,7 @@
   volume={13},
   number={2},
   pages={31--39},
+  doi={10.1109/MCSE.2010.118},
   year={2011},
   publisher={IEEE Computer Society}
 }
@@ -80,6 +82,7 @@
   volume={13},
   number={2},
   pages={22--30},
+  doi={10.1109/MCSE.2011.37},
   year={2011},
   publisher={IEEE Computer Society}
 }
@@ -89,6 +92,7 @@
   author={Aggarwal, Charu C and others},
   volume={1},
   year={2016},
+  doi={10.1007/978-3-319-29659-3},
   publisher={Springer}
 }
 
