### 2020-07-24 #2174 @NicolasHug
@terrytangyuan version that I published on zenodo is 1.1.1, not 1.1.0 (not sure if that matters?)
