### 2020-07-24 #2174 @terrytangyuan
@NicolasHug Could you move the paper to master branch?
