### 2020-07-24 #2174 @NicolasHug
Isn't there a way to accept the paper from another branch? Ideally I'd rather keep the master branch away from non-code related files.
