### 2020-07-25 #2174 @Kevin-Mattheus-Moerman
@NicolasHug these are some additional comments on your paper:

- [ ] Please make the affiliation complete and make sure it includes department, city, and country (write United States of America in full).
- [ ] Remove additional space after bracket last bracket in `(Buitinck et al., 2013; Pedregosa et al., 2011) , ...`

Once you made these changes use `@whedon generate pdf` here to update the paper.
