### 2020-07-27 #2174 @NicolasHug
I have removed the space, thanks.

Regarding the affiliation, would you have more detailed guidelines? The current version seems to be compliant with the [instructions](https://joss.readthedocs.io/en/latest/submitting.html), judging by the provided example.
