### 2020-08-02 #2174 @Kevin-Mattheus-Moerman
> I have removed the space, thanks.
>
> Regarding the affiliation, would you have more detailed guidelines? The current version seems to be compliant with the [instructions](https://joss.readthedocs.io/en/latest/submitting.html), judging by the provided example.

I think our guidelines are not clear enough on this, and we are updating these at the moment. If you could change it to have: `Institution, department (if applicable), city, state, country` that would be great. And please spell out United States of America.
Thanks.
