### 2020-08-02 #2174 @NicolasHug
Thanks @Kevin-Mattheus-Moerman , I updated to

`- name: Columbia University, Data Science Institute, New York City, New York, United States of America`
