### 2020-08-05 #2174 @danielskatz
@NicolasHug - in the references, I think "ECML pkdd workshop" should be "ECML PKDD Workshop" Similarly the booktitle for Gantner should be changed to title case (by protecting it in the .bib file with {}s
Guo is missing something
Some caps need to be fixed in Harper
Same for Kula 2015
And Lee
And McKinney
and Yang

After these are fixed, please regenerate the pdf
