+++ paper.bib
@@ -18,7 +18,7 @@
                Arnaud Joly and Brian Holt and Ga{\"{e}}l Varoquaux},
   title     = {{API} design for machine learning software: experiences from the scikit-learn
                project},
-  booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
+  booktitle = {{ECML PKDD Workshop: Languages for Data Mining and Machine Learning}},
   year      = {2013},
   pages = {108--122},
 }
@@ -26,7 +26,7 @@
 @inproceedings{NMF,
   title={Algorithms for non-negative matrix factorization},
   author={Lee, Daniel D and Seung, H Sebastian},
-  booktitle={Advances in neural information processing systems},
+  booktitle={{Advances in Neural Information Processing Systems}},
   pages={556--562},
   year={2001}
 }
@@ -43,9 +43,9 @@
 }
 
 @article{movielens,
-  title={The movielens datasets: History and context},
+  title={The Movielens Datasets: History and Context},
   author={Harper, F Maxwell and Konstan, Joseph A},
-  journal={Acm transactions on interactive intelligent systems (tiis)},
+  journal={ACM Transactions on Interactive Intelligent Systems (TIIS)},
   volume={5},
   number={4},
   pages={1--19},
@@ -55,8 +55,8 @@
 
 @InProceedings{pandas,
   author    = { Wes McKinney },
-  title     = { Data Structures for Statistical Computing in Python },
-  booktitle = { Proceedings of the 9th Python in Science Conference },
+  title     = { {Data Structures for Statistical Computing in Python} },
+  booktitle = { {Proceedings of the 9th Python in Science Conference}},
   pages     = { 51 - 56 },
   doi       = {10.25080/Majora-92bf1922-00a},
   year      = { 2010 },
@@ -96,14 +96,14 @@
   publisher={Springer}
 }
 
-@inproceedings{librec,
+@misc{librec,
   title={LibRec: A Java Library for Recommender Systems.},
   author={Guo, Guibing and Zhang, Jie and Sun, Zhu and Yorke-Smith, Neil}
 }
 
 @inproceedings{mymedialite,
     author    = { Zeno Gantner and Steffen Rendle and Christoph Freudenthaler and Lars Schmidt-Thieme },
-    title     = { {MyMediaLite}: A Free Recommender System Library },
+    title     = { {MyMediaLite: A Free Recommender System Library }},
     booktitle = { Proceedings of the 5th ACM Conference
                   on Recommender Systems (RecSys 2011) },
     year      = 2011
@@ -119,14 +119,14 @@
 
 @inproceedings{openrec,
 author = {Yang, Longqi and Bagdasaryan, Eugene and Gruenstein, Joshua and Hsieh, Cheng-Kang and Estrin, Deborah},
-title = {OpenRec: A Modular Framework for Extensible and Adaptable Recommendation Algorithms},
+title = {{OpenRec: A Modular Framework for Extensible and Adaptable Recommendation Algorithms}},
 year = {2018},
 isbn = {9781450355810},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
 url = {https://doi.org/10.1145/3159652.3159681},
 doi = {10.1145/3159652.3159681},
-booktitle = {Proceedings of the Eleventh ACM International Conference on Web Search and Data Mining},
+booktitle = {{Proceedings of the Eleventh ACM International Conference on Web Search and Data Mining}},
 pages = {664–672},
 numpages = {9},
 keywords = {recommendation, adaptable, modular, extensible, framework},
@@ -139,9 +139,9 @@ series = {WSDM ’18}
   editor    = {Toine Bogers and
                Marijn Koolen},
[…]
