### 2020-08-05 f949d3c
some more ref fixes
