+++ paper.bib
@@ -1,5 +1,5 @@
 @article{scikit-learn,
- title={Scikit-learn: Machine Learning in {P}ython},
+ title={Scikit-learn: Machine {L}earning in {P}ython},
  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
          and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
          and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
@@ -43,7 +43,7 @@
 }
 
 @article{movielens,
-  title={The Movielens Datasets: History and Context},
+  title={{The Movielens Datasets: History and Context}},
   author={Harper, F Maxwell and Konstan, Joseph A},
   journal={ACM Transactions on Interactive Intelligent Systems (TIIS)},
   volume={5},
@@ -97,15 +97,15 @@
 }
 
 @misc{librec,
-  title={LibRec: A Java Library for Recommender Systems.},
+  title={{LibRec: A Java Library for Recommender Systems}},
   author={Guo, Guibing and Zhang, Jie and Sun, Zhu and Yorke-Smith, Neil}
 }
 
 @inproceedings{mymedialite,
     author    = { Zeno Gantner and Steffen Rendle and Christoph Freudenthaler and Lars Schmidt-Thieme },
     title     = { {MyMediaLite: A Free Recommender System Library }},
-    booktitle = { Proceedings of the 5th ACM Conference
-                  on Recommender Systems (RecSys 2011) },
+    booktitle = {{ Proceedings of the 5th ACM Conference
+                  on Recommender Systems (RecSys 2011) }},
     year      = 2011
 }
 
@@ -138,7 +138,7 @@ series = {WSDM ’18}
   author    = {Maciej Kula},
   editor    = {Toine Bogers and
                Marijn Koolen},
-  title     = {Metadata Embeddings for User and Item Cold-start Recommendations},
+  title     = {{Metadata Embeddings for User and Item Cold-start Recommendations}},
   booktitle = {{Proceedings of the 2nd Workshop on New Trends on Content-Based Recommender
                Systems co-located with 9th ACM Conference on Recommender Systems
                (RecSys 2015), Vienna, Austria, September 16-20, 2015.}},
