### 2020-08-05 #2174 @NicolasHug
Thanks @danielskatz , capitalization is hopefully fixed now
